Separate gift-links types into database, queries and models modules
Splits the single model.ts into database.ts (table schemas + knex types),
queries.ts (read projection, the row->domain codec, and query statements),
and models.ts (the domain types, no codecs). Each row shape still has one
source: the projection and knex types both derive from the DbGiftLink schema.
